+++ src/main/java/com/homeheaven/backend/repository/ProductsRepository.java
@@ -7,12 +7,13 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductsRepository extends JpaRepository<Product, Integer> {
+public interface ProductsRepository extends JpaRepository<Product, Long> {
 
     @Query("SELECT p FROM Product p WHERE p.category = ?1")
     Iterable<Product> getProductsByCategory(String category);
 
-
     @Query("SELECT p FROM Product p WHERE p.productName LIKE %?1%")
     Iterable<Product> searchProducts(String param);
+
+
 }
